pyproject.toml
[project]
name = "pedantic-lawyer"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "colorlog>=6.9.0",
    "diff-match-patch>=20241021",
    "markitdown-rtf-plugin>=0.1.0",
    "markitdown[all]>=0.1.1",
    "natasha>=1.6.0",
    "openai>=1.69.0",
    "pydantic>=2.11.0",
    "pymorphy2-dicts-ru>=2.4.417127.4579844",
    "pytest>=8.3.5",
    "pytest-asyncio>=0.26.0",
    "pyyaml>=6.0.2",
    "setuptools>=78.1.0",
    "streamlit>=1.44.0",
    "tabulate>=0.9.0",
    "types-pyyaml>=6.0.12.20250326",
]

[tool.ruff]
target-version = "py312"
line-length = 120
lint.ignore = ["PLR"]
lint.extend-select = ["I", "UP", "PL", "B", "Q"]
